Add library/term/mode defaults to inputrc parsing
